refactor(dedup): A4 catchup/impact → shared runAgentCli dispatcher
Extract the byte-identical agent-CLI dispatcher body (gateway-state→IPC connect→awaitAgentBrief+timeout→call→renderAgentBrief→exit-code error handling→finally disconnect) into cli/src/lib/agent-cli-dispatcher.ts; catchup + impact now build callParams and delegate. Exact stderr text + exit(1|2) codes preserved. Both command test suites (31) pass unedited.
