+++ packages/cli/src/commands/impact.ts
@@ -1,8 +1,4 @@
-import { IPCClient } from "../ipc-client/index.ts";
-import { awaitAgentBrief, renderAgentBrief } from "../lib/agent-brief-render.ts";
-import { readGatewayState } from "../lib/gateway-process.ts";
-import { registerInteractiveCliIpcHandlers } from "../lib/interactive-ipc-handlers.ts";
-import { getCliPlatformPaths } from "../paths.ts";
+import { runAgentCli } from "../lib/agent-cli-dispatcher.ts";
 import { isImpactBrief } from "../types/agents.ts";
 
 export type ImpactCliArgs = {
@@ -60,38 +56,16 @@ export function parseImpactArgs(args: string[]): ImpactCliArgs {
 
 export async function runImpactCli(args: string[]): Promise<void> {
   const parsed = parseImpactArgs(args);
-
-  const paths = getCliPlatformPaths();
-  const state = await readGatewayState(paths);
-  if (state === undefined) {
-    process.stderr.write("Gateway is not running. Start with: nimbus start\n");
-    process.exit(1);
-  }
-
-  const client = new IPCClient(state.socketPath);
-  await client.connect();
-  registerInteractiveCliIpcHandlers(client);
-
-  let timeout: ReturnType<typeof setTimeout> | undefined;
-  const briefPromise = awaitAgentBrief(client, "impact", isImpactBrief, (t) => {
-    timeout = t;
-  });
-
   const callParams: { fileOrPrUrl: string; depth?: number; service?: string } = {
     fileOrPrUrl: parsed.fileOrPrUrl,
   };
   if (parsed.depth !== undefined) callParams.depth = parsed.depth;
   if (parsed.service !== undefined) callParams.service = parsed.service;
-
-  try {
-    await client.call<{ sessionId: string }>("agents.impact", callParams);
-    const { brief, findings } = await briefPromise;
-    renderAgentBrief(brief, findings, parsed.json);
-  } catch (err) {
-    process.stderr.write(`${err instanceof Error ? err.message : String(err)}\n`);
-    process.exit(2);
-  } finally {
-    if (timeout !== undefined) clearTimeout(timeout);
-    await client.disconnect();
-  }
+  await runAgentCli({
+    agentName: "impact",
+    ipcMethod: "agents.impact",
+    callParams,
+    guard: isImpactBrief,
+    json: parsed.json,
+  });
 }
