+++ packages/cli/src/lib/agent-cli-dispatcher.ts
@@ -1,0 +1,50 @@
+import { IPCClient } from "../ipc-client/index.ts";
+import { getCliPlatformPaths } from "../paths.ts";
+import { awaitAgentBrief, renderAgentBrief } from "./agent-brief-render.ts";
+import { readGatewayState } from "./gateway-process.ts";
+import { registerInteractiveCliIpcHandlers } from "./interactive-ipc-handlers.ts";
+
+/**
+ * Run an agent CLI command: read the gateway state (exit 1 if not running),
+ * connect the IPC client, await the agent's brief (guarded by `guard`), invoke
+ * `ipcMethod` with `callParams`, then render the brief. On error: stderr + exit 2.
+ * Always clears the brief timeout + disconnects in `finally`. The per-command
+ * code supplies only the agent name, IPC method, call params, brief guard, and
+ * the `--json` flag — this collapses the byte-identical dispatcher body shared by
+ * the agent commands (catchup, impact, …).
+ */
+export async function runAgentCli<B extends { gaps: readonly { category: string }[] }>(opts: {
+  agentName: string;
+  ipcMethod: string;
+  callParams: Record<string, unknown>;
+  guard: (x: unknown) => x is B;
+  json: boolean;
+}): Promise<void> {
+  const paths = getCliPlatformPaths();
+  const state = await readGatewayState(paths);
+  if (state === undefined) {
+    process.stderr.write("Gateway is not running. Start with: nimbus start\n");
+    process.exit(1);
+  }
+
+  const client = new IPCClient(state.socketPath);
+  await client.connect();
+  registerInteractiveCliIpcHandlers(client);
+
+  let timeout: ReturnType<typeof setTimeout> | undefined;
+  const briefPromise = awaitAgentBrief(client, opts.agentName, opts.guard, (t) => {
+    timeout = t;
+  });
+
+  try {
+    await client.call<{ sessionId: string }>(opts.ipcMethod, opts.callParams);
+    const { brief, findings } = await briefPromise;
+    renderAgentBrief(brief, findings, opts.json);
+  } catch (err) {
+    process.stderr.write(`${err instanceof Error ? err.message : String(err)}\n`);
+    process.exit(2);
+  } finally {
+    if (timeout !== undefined) clearTimeout(timeout);
+    await client.disconnect();
+  }
+}
